+++ libs/data_loaders/subwords.py
@@ -54,7 +54,7 @@ def create_tf_encoder_bilingual(tokenizer1, tokenizer2):
 
     def tf_py_encode(lang1, lang2):
         result_lang1, result_lang2 = tf.py_function(
-            encode, [lang1, lang2], [tf.int64, tf.int64])
+            encode, [lang1, lang2], [tf.int32, tf.int32])
         result_lang1.set_shape([None])
         result_lang2.set_shape([None])
