+++ libs/models/transformerv2.py
@@ -30,9 +30,9 @@ def builder(config: typing.Dict[typing.AnyStr, typing.Any]):
     final_layer = tf.keras.layers.Dense(vocab_size_target)
 
     enc_inp = tf.keras.layers.Input(
-        shape=(None,), dtype=tf.int64, name="i1")
+        shape=(None,), dtype=tf.int32, name="i1")
     dec_inp = tf.keras.layers.Input(
-        shape=(None,), dtype=tf.int64, name="i2")
+        shape=(None,), dtype=tf.int32, name="i2")
     padding_mask = tf.keras.layers.Input(
         shape=(1, 1, None), dtype=tf.float32, name="i3")
     combined_mask = tf.keras.layers.Input(
